fix bug

+++ core.py
@@ -1,1 +1,1 @@
-def add(a, b): return float(a) + float(b)
+def add(a, b): return int(a) + int(b)
